requirements.txt
fastapi
pydantic
uvicorn
py-grpc-prometheus
python-logging-loki
psutil